Move allow ip list to prop. The intention is to make it externalized. Also moved config classes under configuration package.
By setting OS env variable SECURITY_ALLOW_IP_LIST to include device ip addresses, the server will accept direct access from such devices.
If a proxy/load balancer is used, its address may be specified.
