+++ src/main/resources/application.properties
@@ -20,3 +20,5 @@ spring.cache.caffeine.spec=initialCapacity=10,maximumSize=50,expireAfterWrite=10
 spring.cache.redis.host=localhost
 spring.cache.redis.port=6379
 spring.cache.redis.time-to-live=60000
+
+security.allow.ip.list=127.0.0.0/8,::1,192.168.1.0/24
